basket created

+++ back/models/models.js
@@ -9,7 +9,7 @@ const User =sequelize.define('user',{
 })
 
 const Basket = sequelize.define('basket',{
-    id: {type: DataTypes.INTEGER, primaryKey:true, autoIncrement:false},
+    id: {type: DataTypes.BIGINT, primaryKey:true, autoIncrement:false},
 })
 const BasketDevice = sequelize.define('basket_device',{
     id: {type: DataTypes.INTEGER, primaryKey:true, autoIncrement:true},
